requirements.txt
datasets
faiss-cpu
sentence-transformers
transformers
torch
numpy
pandas
rouge-score
mauve-text
tqdm
pytest
langchain
langchain-community
pypdf
langchain-huggingface